+++ RELEASENOTES.md
@@ -9,6 +9,8 @@
 - `ApiUser` deprecated in favour of `UserAndPassword` and `UserAndPasswordWithDomain`
 - `MediawikiApiInterface` `isLoggedin` renamed to `isLoggedIn`
 - `MediawikiApiInterface` and implementations will now only return a boolean from `isLoggedIn`
+- `MediawikiApi` constructor signature changed
+- `SimpleRequest` now extends `FluentRequest`
 
 ## Version 2.8 (16 February 2021)